requirements.txt
coqui-tts
torch
python-multipart
fastapi[standard]
uvicorn